Refactor to a separate object
